feat(sign): keep import and login in sign page dialog
Extract ImportForm for reuse, switch login/import inside the sign auth
dialog without leaving the page, and log in immediately after import.
