+++ src/i18n/translations.json
@@ -90,7 +90,8 @@
       "callbackNotice": "You will be redirected after signing.",
       "authorityNotice": "This transaction requires your {{authority}} key.",
       "continueToLogin": "Continue to login",
-      "loginDialogDescription": "Log in to review and sign this transaction."
+      "loginDialogDescription": "Log in to review and sign this transaction.",
+      "importDialogDescription": "Import your account to review and sign this transaction."
     },
     "dashboard": {
       "title": "Dashboard",
@@ -249,7 +250,8 @@
       "callbackNotice": "签名后您将被重定向。",
       "authorityNotice": "此交易需要您的 {{authority}} 密钥。",
       "continueToLogin": "继续登录",
-      "loginDialogDescription": "登录后即可查看并签名此交易。"
+      "loginDialogDescription": "登录后即可查看并签名此交易。",
+      "importDialogDescription": "导入账号后即可查看并签名此交易。"
     },
     "dashboard": {
       "title": "控制台",
